chore: update RMCP version to latest (#328)
This commit updates the RMCP dependency to the latest version 0.6.4.
Sadly, schemars was also updated with this, so a lot of unrelated
changes were needed to conform with the new stable schemars version.
