+++ .changesets/maint_nc_rmcp_update.md
@@ -1,0 +1,3 @@
+### chore: update RMCP dependency ([328](https://github.com/apollographql/apollo-mcp-server/issues/328))
+
+Update the RMCP dependency to the latest version, pulling in newer specification changes.
