+++ crates/apollo-mcp-server/Cargo.toml
@@ -24,12 +24,13 @@ futures.workspace = true
 headers = "0.4.1"
 http = "1.3.1"
 humantime-serde = "1.1.1"
+jsonschema = "0.33.0"
 jsonwebtoken = "9"
 jwks = "0.4.0"
 lz-str = "0.2.1"
 regex = "1.11.1"
 reqwest.workspace = true
-rmcp = { version = "0.2", features = [
+rmcp = { version = "0.6", features = [
   "server",
   "transport-io",
   "transport-sse-server",
