+++ crates/apollo-mcp-server/src/explorer.rs
@@ -82,8 +82,10 @@ impl Explorer {
         let url = self.create_explorer_url(input)?;
         debug!(?url, input=?pretty, "Created URL to open operation in Apollo Explorer");
         Ok(CallToolResult {
-            content: vec![Content::text(url)],
+            content: vec![Content::text(url.clone())],
+            meta: None,
             is_error: None,
+            structured_content: Some(Value::Array(vec![url.into()])),
         })
     }
 }
