+++ crates/apollo-mcp-server/src/introspection/tools/introspect.rs
@@ -73,6 +73,8 @@ impl Introspect {
                 return Ok(CallToolResult {
                     content: vec![],
                     is_error: None,
+                    meta: None,
+                    structured_content: None,
                 });
             }
         }
@@ -99,6 +101,9 @@ impl Introspect {
                 .map(Content::text)
                 .collect(),
             is_error: None,
+            meta: None,
+            // The content being returned is a raw string, so no need to create structured content for it
+            structured_content: None,
         })
     }
