+++ crates/apollo-mcp-server/src/introspection/tools/validate.rs
@@ -70,6 +70,11 @@ impl Validate {
         Ok(CallToolResult {
             content: vec![Content::text("Operation is valid")],
             is_error: None,
+            meta: None,
+
+            // Note: We don't really return any meaningful content to the client here, so we can leave the
+            // structured content as none.
+            structured_content: None,
         })
     }
 }
